\name{AIC.modgam}
\alias{AIC.modgam}
\title{
AIC of the \code{\link{modgam}} Object}
\description{
This function summarizes the AIC of a modgam object produced by \code{\link{modgam}}.
}
\usage{
\method{AIC}{modgam} (object, \dots) 
}
\arguments{
  \item{object}{
a modgam object.
}
  \item{\dots}{
extra arguments for S3 generic, ignored by \code{AIC.modgam}.   	
}
}
\author{
Lu Bai

Send bug reports to \email{sbartell@uci.edu}.
}
\seealso{
\code{\link{modgam}}
\code{\link{gamcox}}, 
\code{\link{predict.gamcox}}.
}
\keyword{misc}
\keyword{smooth}
