\name{formula.modgam}
\alias{formula.modgam}
\title{
Formula of the \code{\link{modgam}} Object}
\description{
This function provide the formula of a modgam object produced by \code{\link{modgam}}.
}
\usage{
\method{formula}{modgam} (x, \dots) 
}
\arguments{
  \item{x}{
a modgam object.
}
  \item{\dots}{
extra arguments for S3 generic, ignored by \code{formula.modgam}.   	
}
}
\author{
Lu Bai

Send bug reports to \email{sbartell@uci.edu}.
}
\seealso{
\code{\link{modgam}}
\code{\link{gamcox}}, 
\code{\link{predict.gamcox}}.
}
\keyword{misc}
\keyword{smooth}
