\name{MapGAM.version}
\alias{MapGAM.version}
\title{
Revision dates for MapGAM functions
}
\description{
Print version and dates for MapGAM package and functions
}
\usage{
  MapGAM.version()
}
\author{
	Scott Bartell	
Send bug reports to \email{sbartell@uci.edu}.
}
\seealso{
\code{\link{MapGAM-package}}, 
}
\keyword{version}
