\name{print.modgam}
\alias{print.modgam}
\title{
Print modgam Objects}
\description{
This function displays information about a modgam object produced by \code{\link{modgam}}.
}
\usage{
\method{print}{modgam} (x, \dots) 
}
\arguments{
  \item{x}{
a modgam object.
}
  \item{\dots}{
extra arguments for S3 generic, ignored by \code{print.modgam}   	
}
}
\author{
Lu Bai

Send bug reports to \email{sbartell@uci.edu}.
}
\seealso{
\code{\link{modgam}}
\code{\link{gamcox}}, 
\code{\link{predict.gamcox}}.
}
\keyword{misc}
\keyword{smooth}
