\name{residuals.modgam}
\alias{residuals.modgam}
\title{
Residuals of the \code{\link{modgam}} Object}
\description{
This function provides residuals of a modgam object produced by \code{\link{modgam}}.
}
\usage{
\method{residuals}{modgam} (object, \dots) 
}
\arguments{
  \item{object}{
a modgam object.
}
  \item{\dots}{
extra arguments for S3 generic, ignored by \code{residuals.modgam}.   	
}
}
\author{
Lu Bai

Send bug reports to \email{sbartell@uci.edu}.
}
\seealso{
\code{\link{modgam}}
\code{\link{gamcox}}, 
\code{\link{predict.gamcox}}.
}
\keyword{misc}
\keyword{smooth}
