\name{summary.modgam}
\alias{summary.modgam}
\title{
Summarize the \code{\link{modgam}} Object}
\description{
This function summarizes information about a modgam object produced by \code{\link{modgam}}.
}
\usage{
\method{summary}{modgam} (object, \dots) 
}
\arguments{
  \item{object}{
a modgam object.
}
  \item{\dots}{
extra arguments for S3 generic, ignored by \code{summary.modgam}.   	
}
}
\author{
Lu Bai

Send bug reports to \email{sbartell@uci.edu}.
}
\seealso{
\code{\link{modgam}}
\code{\link{gamcox}}, 
\code{\link{predict.gamcox}}.
}
\keyword{misc}
\keyword{smooth}
