\name{MapGAM.version}
\alias{MapGAM.version}
\title{
Revision Dates for MapGAM Functions
}
\description{
Print version and dates for MapGAM package and functions.
}
\usage{
  MapGAM.version()
}
\author{
	Scott Bartell	
Send bug reports to \email{sbartell@uci.edu}.
}
\seealso{
\code{\link{MapGAM-package}}, 
}
\keyword{version}
